image upload fixed for testing

+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/En/ItemDto.java
@@ -1,6 +1,6 @@
 package hu.project.groupproject.resourceserver.dtos.En;
 
-public record ItemDto(String userId, String name, String description, String state, String location, Integer phone) {}
+public record ItemDto(String userId, String name, String description, String state, String location, Long phone) {}
 // "userId":string
 // "nev": string
 // "leiras": string
