+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/En/NoticeDto.java
@@ -2,7 +2,7 @@ package hu.project.groupproject.resourceserver.dtos.En;
 
 import java.sql.Timestamp;
 
-public record NoticeDto(String userId,String type, String urgency, String description, String location, Integer phone, Timestamp date) {}
+public record NoticeDto(String userId,String type, String urgency, String description, String location, Long phone, Timestamp date) {}
 // "userid":string
 // "tipus": string
 // "sulyossag": string
