+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/En/ReservationDto.java
@@ -2,7 +2,7 @@ package hu.project.groupproject.resourceserver.dtos.En;
 
 import java.sql.Timestamp;
 
-public record ReservationDto(String userId, String orgId, String preferredName, String email, Integer phone, Timestamp startDate,Timestamp endDate) {}
+public record ReservationDto(String userId, String orgId, String preferredName, String email, Long phone, Timestamp startDate,Timestamp endDate) {}
 // "userId":string
 // "orgId":string
 // "nev": string
