+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/En/UserInfoDto.java
@@ -2,7 +2,7 @@ package hu.project.groupproject.resourceserver.dtos.En;
 
 import java.util.Set;
 
-public record UserInfoDto(String userId, String email, String name, Integer phone, Set<String> roles,String location, Set<String> orgs) {}
+public record UserInfoDto(String userId, String email, String name, Long phone, Set<String> roles,String location, Set<String> orgs) {}
 // "userId":string
 // "email":string
 // "nev": string
