+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/En/users/UserDtoNew.java
@@ -1,3 +1,3 @@
 package hu.project.groupproject.resourceserver.dtos.En.users;
 
-public record UserDtoNew(String userName, String firstName, String lastName, Integer phone, String email) {}
+public record UserDtoNew(String userName, String firstName, String lastName, Long phone, String email) {}
