+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/Hu/EventDtoHu.java
@@ -3,7 +3,7 @@ package hu.project.groupproject.resourceserver.dtos.Hu;
 import java.util.List;
 import java.sql.Timestamp;
 
-public record EventDtoHu(String nev, String leiras, String helyszin, Long userId, Long orgId, List<Integer> telefonszam, List<String> email, Timestamp startDate, Timestamp endDate) {}
+public record EventDtoHu(String nev, String leiras, String helyszin, Long userId, Long orgId, List<Long> telefonszam, List<String> email, Timestamp startDate, Timestamp endDate) {}
 // "nev": string //rendezvény neve
 // "leiras": string
 // "helyszin": string
