+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/Hu/ItemDtoHu.java
@@ -1,6 +1,6 @@
 package hu.project.groupproject.resourceserver.dtos.Hu;
 
-public record ItemDtoHu(String userId, String nev, String leiras, String allapot, String helyszin, Integer telefonszam) {}
+public record ItemDtoHu(String userId, String nev, String leiras, String allapot, String helyszin, Long telefonszam) {}
 // "userId":string
 // "nev": string
 // "leiras": string
