+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/Hu/NoticeDtoHu.java
@@ -2,7 +2,7 @@ package hu.project.groupproject.resourceserver.dtos.Hu;
 
 import java.sql.Timestamp;
 
-public record NoticeDtoHu(String userId,String tipus, String sulyossag, String leiras, String helyszin, Integer telefonszam, Timestamp date) {}
+public record NoticeDtoHu(String userId,String tipus, String sulyossag, String leiras, String helyszin, Long telefonszam, Timestamp date) {}
 // "userid":string
 // "tipus": string
 // "sulyossag": string
