+++ resource-server/src/main/java/hu/project/groupproject/resourceserver/dtos/Hu/UserInfoDtoHu.java
@@ -2,7 +2,7 @@ package hu.project.groupproject.resourceserver.dtos.Hu;
 
 import java.util.Set;
 
-public record UserInfoDtoHu(String userId, String email, String nev, Integer telefonszam, Set<String> roles,String helyszin, Set<String> orgs) {}
+public record UserInfoDtoHu(String userId, String email, String nev, Long telefonszam, Set<String> roles,String helyszin, Set<String> orgs) {}
 // "userId":string
 // "email":string
 // "nev": string
